fix 2239 remove build warnings
This commit removes 150+ warnings. Most of them are fixed by using
the diamond operator where it hadn't been used previously.
Some fixes involve actually "fixing" compile type problems (that weren't
problematic in practice in tests due to erasure).
Also fixes warnings due to the absence of findbugs annotations.jar on
the compile classpath of reactor-tools (needed when bytebuddy runs).

+++ reactor-core/src/test/java/reactor/core/publisher/FluxBufferPredicateTest.java
@@ -979,7 +979,7 @@ public class FluxBufferPredicateTest {
 	public void scanOperator(){
 		Flux<Integer> parent = Flux.just(1);
 		FluxBufferPredicate<Integer, ArrayList<Integer>> test =
-				new FluxBufferPredicate<>(parent, v -> (Integer) v != 0, ArrayList::new, FluxBufferPredicate.Mode.UNTIL);
+				new FluxBufferPredicate<>(parent, v -> v != 0, ArrayList::new, FluxBufferPredicate.Mode.UNTIL);
 
 		assertThat(test.scan(Scannable.Attr.PARENT)).isSameAs(parent);
 		assertThat(test.scan(Scannable.Attr.RUN_STYLE)).isSameAs(Scannable.Attr.RunStyle.SYNC);
